Remove old spotifyPlayer, rename the new
