+++ conf.json.example
@@ -26,8 +26,6 @@
   },
   "SpotifyPlayer": {
      "Enabled": true,
-     "Host": "127.0.0.1",
-     "UseConnect": false,
      "ClientID": "<clientID>",
      "ClientSecret": "<secret>",
      "TokenFilePath": ".musicbot-spotify-token"
